Updates to address issues in section 2 + minor cleanup
I found some areas where the tasks didn’t line up with Guide, and some unused variables.

The biggest problem was most of section 2.1 - this was trying to handle xinetd services as system services, which isn’t correct. The way to enable/disable these would be to find files with the service name, and make sure that the line disable = yes exists in each file.

I might be making this change locally, but for now, I just “unimplemented” all of them to make it clear that they won’t work as intended, since in the default case, we would delete xinetd right after these steps.